refactor #120 : 가독성을 위해 공통 레이아웃에 사용되는 버튼명 통일
